gptel-agent-tools: Fix task overlay formatting
* gptel-agent-tools.el (gptel-agent--task-overlay): Because we now
use buffer text to show pending tool calls, the agent_task overlay
can end up on the tool confirmation text instead of the LLM
response or user prompt text.  This means the overlay evaporates
when the tool confirmation prompt goes away.  Fix by making the
task overlay not evaporative.

It is cleaned up in the agent_task callback, but there might be
edge cases of request failures where it lingers.  This needs to be
investigated further.

An alternative solution is to place the task overlay over the same
text as the tool confirmation overlay, i.e. everything from
:position to :tracking-marker.